Backend/requirements.txt
django
djangorestframework
environs
asgiref
Django
django-cors-headers
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
python-dateutil
channels
channels-redis
daphne
